Branched from master
Created a Node class which now can contain a Box and a Goal

Changed addGoals to add Nodes instead of Goals

Changed isGoalState to check each Node containing a Goal, whether or
not it contains a box - if all checks goes through, isGoalState will
return true and a solution has been found.

Currently, this branch does not find a solution.